+++ src/engine/util/log.rs
@@ -8,16 +8,24 @@ pub fn set_logging_enabled(is_enabled: bool) {
     }
 }
 
-// FIXME: It's not ideal to open a handle to the file every time we want to write a line
+pub fn crashlog<S: AsRef<str>>(s: S) {
+    log_to_file(s, "crash.log");
+}
+
 pub fn log<S: AsRef<str>>(s: S) {
     if unsafe { !ENABLE_LOGGING } {
         return;
     }
 
+    log_to_file(s, "log");
+}
+
+// FIXME: It's not ideal to open a handle to the file every time we want to write a line
+fn log_to_file<S: AsRef<str>>(s: S, extension: &str) {
     let current_exe =
         std::env::current_exe().expect("Unable to determine current executable directory");
 
-    let path = current_exe.with_extension("log");
+    let path = current_exe.with_extension(extension);
 
     let mut f = fs::OpenOptions::new()
         .create(true)