+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 * Bundle midgame and endgame evals into a `PhasedEval` struct
 * Fix taking up more memory than needed when the transposition table is resized repeatedly
 * Remove the default 50ms move overhead and add a UCI option to configure it
+* Always log crashes to a .crash.log file
 
 ## [2.1]